Organize Grammer member files.
