+++ src/Parsley/Grammar.Pattern.cs
